+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - converter names as services IDs
+
+### Deprecated
+ - converter service IDs as `aymdev_commonmark.converter.CONVERTER_NAME` 
 
 ## [1.2.0] - 2020-10-25
 ### Added