+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.1] - 2020-11-08
+### Fixed
+ - definition deprecation for *old* service IDs has been corrected for Symfony < 5.1
+
 ## [1.3.0] - 2020-11-07
 ### Added
  - converter names as services IDs